+++ dev/src/route.js
@@ -6,7 +6,8 @@
      * @constructor
      */
     function Route(pattern, callback, priority, router) {
-        var isRegexPattern = isRegExp(pattern);
+        var isRegexPattern = isRegExp(pattern),
+            patternLexer = crossroads.patternLexer;
         this._router = router;
         this._pattern = pattern;
         this._paramsIds = isRegexPattern? null : patternLexer.getParamIds(this._pattern);
@@ -66,7 +67,7 @@
 
         _getParamsObject : function (request) {
             var shouldTypecast = this._router.shouldTypecast,
-                values = patternLexer.getParamValues(request, this._matchRegexp, shouldTypecast),
+                values = crossroads.patternLexer.getParamValues(request, this._matchRegexp, shouldTypecast),
                 o = {},
                 n = values.length;
             while (n--) {
@@ -87,7 +88,7 @@
             if (norm && isFunction(norm)) {
                 params = norm(request, this._getParamsObject(request));
             } else {
-                params = patternLexer.getParamValues(request, this._matchRegexp, this._router.shouldTypecast);
+                params = crossroads.patternLexer.getParamValues(request, this._matchRegexp, this._router.shouldTypecast);
             }
             return params;
         },